requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
future==1.0.0
Google-Search-API==1.1.14
google_search_results==2.4.2
h11==0.16.0
idna==3.10
multidict==6.6.3
numpy==2.2.6
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.3.1
propcache==0.3.2
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
selenium==2.53.6
serpapi==0.1.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
vcrpy==7.0.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
